Linux

+++ Smithtry1000/Line.h
@@ -4,11 +4,11 @@
 class LinesElement :public VerticalLinesElement
 {
 private:
-	double _alpha;
+	long double _alpha;
 
 public:
-	void SetAlpha(double);
-	double GetAlpha();
-	LinesElement(mode mode, double value, double frequency, Point point, map<chartMode, tuple<double, double>> chartParameters,
-		map<parameterMode, Complex> parameters, double electricalLength, double mechanicalLength, double theta, double lambda, double alpha);
+	void SetAlpha(long double);
+	long double GetAlpha();
+	LinesElement(mode mode, long double value, long double frequency, Point point, map<chartMode, tuple<long double, long double>> chartParameters,
+		map<parameterMode, Complex> parameters, long double electricalLength, long double mechanicalLength, long double theta, long double lambda, long double alpha);
 };
